move archetype fragmentation benchmarks to a separate file
